vyatta-config-mgmt: config archive fails on unknown ssh host key
If the remote host configured for commit-archive hasn't been connected
to before then it's host key is missing in known_hosts.  If this is the
case, then prompt the user to confirm the hosts authenticity after
presenting them with the host keys fingerprint, inserting it into
known_hosts and continuing with the archive if it's correct.

Bug #517 http://bugzilla.vyos.net/show_bug.cgi?id=517